Reorganized the files into different projects
