+++ civic-tech-club/src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Home from './pages/Home';  // Your existing Home component
-import LondonBudget from './pages/LondonBudget'; // The new About component
-import PDCalendar from './pages/PDCalendar';
+import Home from './pages/Home/Home';  // Your existing Home component
+import LondonBudget from './pages/Budgets/LondonBudget'; // The new About component
+import PDCalendar from './pages/PDCalendar/PDCalendar';
 function App() {
   return (
     <Router>
